+++ pkg/detectors/github_oauth2/github_oauth2.go
@@ -19,6 +19,9 @@ type Scanner struct {
 
 // Ensure the Scanner satisfies the interfaces at compile time.
 var _ detectors.Detector = (*Scanner)(nil)
+var _ detectors.EndpointCustomizer = (*Scanner)(nil)
+
+func (Scanner) CloudEndpoint() string { return "https://api.github.com" }
 
 var (
 	// Oauth2 client ID and secret
